refactor(cli): extract prompt session and image message helpers
